PSYNEU-50 chore: Split sidebar components into multiple files
